// clientproject/src/app/users/models/user.model.ts
export interface User {
  id?: string;
  firstName?: string;
  midName?: string;
  lastName?: string;
  pictureUrl?: string;
  isAdmin?: boolean;
  homePhone?: string;
  gender?: string;
  designation?: string;
  birthDay?: string;
  birthMonth?: string;
  birthYear?: string;
  email?: string;
  phoneNumber?: string;
}